requirements.txt
Flask==1.1.2
pytest==5.3.2; python_version > "3.0"
pytest==5.3.2; python_version < "3.0"
gunicorn==20.0.4
google-cloud-storage==1.32.0
cloudevents
